+++ ChartboostMediationExample/build.gradle
@@ -1,0 +1,67 @@
+apply plugin: 'com.android.application'
+apply plugin: 'kotlin-android'
+
+android {
+    namespace "com.chartboost.mediation.sdk.sampleapp"
+    compileSdkVersion 33
+    buildToolsVersion '33.0.0'
+
+    defaultConfig {
+        multiDexEnabled true
+        applicationId "com.chartboost.mediation.sdk.sampleapp"
+        minSdkVersion 21
+        targetSdkVersion 33
+        versionCode 1
+        versionName "1.0"
+
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+    }
+
+    buildTypes {
+        release {
+            minifyEnabled false
+            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+        }
+    }
+
+    viewBinding {
+        enabled = true
+    }
+}
+
+dependencies {
+    // Chartboost Mediation
+    implementation 'com.chartboost:chartboost-mediation-sdk:4.0.0'
+
+    // Chartboost Mediation requirement
+    implementation 'org.greenrobot:eventbus:3.3.1'
+
+    // Chartboost Mediation Adapters
+    implementation 'com.chartboost:chartboost-mediation-adapter-adcolony:4.4.8.0.0'
+    implementation 'com.chartboost:chartboost-mediation-adapter-admob:4.21.3.0.0'
+    implementation 'com.chartboost:chartboost-mediation-adapter-applovin:4.11.5.5.0'
+    implementation 'com.chartboost:chartboost-mediation-adapter-chartboost:4.9.2.0.0'
+    implementation 'com.chartboost:chartboost-mediation-adapter-amazon-publisher-services:4.9.6.2.0'
+    implementation 'com.chartboost:chartboost-mediation-adapter-digital-turbine-exchange:4.8.2.1.0'
+    implementation 'com.chartboost:chartboost-mediation-adapter-google-bidding:4.21.3.0.0'
+    implementation 'com.chartboost:chartboost-mediation-adapter-inmobi:4.10.1.1.0'
+    implementation 'com.chartboost:chartboost-mediation-adapter-ironsource:4.7.2.5.0.0'
+    implementation 'com.chartboost:chartboost-mediation-adapter-meta-audience-network:4.6.12.0.0'
+    implementation 'com.chartboost:chartboost-mediation-adapter-mintegral:4.16.0.31.0'
+    implementation 'com.chartboost:chartboost-mediation-adapter-tapjoy:4.12.9.1.0'
+    implementation 'com.chartboost:chartboost-mediation-adapter-pangle:4.4.3.0.4.0'
+    implementation 'com.chartboost:chartboost-mediation-adapter-reference:4.1.0.0.0'
+    implementation 'com.chartboost:chartboost-mediation-adapter-unity-ads:4.4.4.1.0'
+    implementation 'com.chartboost:chartboost-mediation-adapter-vungle:4.6.12.0.0'
+    implementation 'com.chartboost:chartboost-mediation-adapter-yahoo:4.1.3.0.0'
+
+    // Google Play Services
+    implementation 'com.google.android.gms:play-services-ads-identifier:18.0.1'
+    implementation 'com.google.android.gms:play-services-base:18.1.0'
+
+    // android app
+    implementation 'com.google.android.material:material:1.8.0'
+    implementation 'androidx.compose.material3:material3:1.1.0-alpha06'
+    implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
+    implementation 'androidx.appcompat:appcompat:1.6.1'
+}
