+++ ChartboostMediationExample/gradle.properties
