+++ ChartboostMediationExample/res/xml/network_security_config.xml
