+++ ChartboostMediationExample/src/main/res/values-v21/theme.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources xmlns:tools="http://schemas.android.com/tools">
+
+    <style name="MainTheme" parent="Theme.Material3.DayNight.NoActionBar" tools:targetApi="lollipop">
+        <item name="android:typeface">normal</item>
+        <item name="android:textSize">15sp</item>
+        <item name="android:textAllCaps">false</item>
+        <item name="android:textColor">#1a222c</item>
+        <item name="colorPrimary">@color/cbGreen80</item>
+
+    </style>
+</resources>
