+++ ChartboostMediationExample/src/main/res/values/colors.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <color name="navy40">#303842</color>
+    <color name="cbGreen80">#5AC648</color>
+    <color name="slate40">#c1d1e0</color>
+</resources>
