+++ ChartboostMediationExample/src/main/res/xml/network_security_config.xml
