--- HeliumExample/build.gradle
@@ -1,117 +1,0 @@
-apply plugin: 'com.android.application'
-apply plugin: 'kotlin-android'
-
-android {
-    namespace "com.chartboost.mediation.sdk.sampleapp"
-    compileSdkVersion 33
-    buildToolsVersion '33.0.0'
-
-    defaultConfig {
-        multiDexEnabled true
-        applicationId "com.chartboost.mediation.example"
-        minSdkVersion 21
-        targetSdkVersion 33
-        versionCode 1
-        versionName "1.0"
-
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-    }
-
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
-        }
-    }
-
-    viewBinding {
-        enabled = true
-    }
-}
-
-dependencies {
-    // Chartboost Mediation
-    implementation 'com.chartboost:chartboost-mediation-sdk:4.0+'
-
-    // Chartboost Mediation requirement
-    implementation 'org.greenrobot:eventbus:3.3.1'
-
-    // Chartboost Mediation Adapters
-    implementation 'com.chartboost:chartboost-mediation-adapter-adcolony:+'
-    implementation 'com.chartboost:chartboost-mediation-adapter-admob:+'
-    implementation 'com.chartboost:chartboost-mediation-adapter-applovin:+'
-    implementation 'com.chartboost:chartboost-mediation-adapter-chartboost:+'
-    implementation 'com.chartboost:chartboost-mediation-adapter-amazon-publisher-services:+'
-    implementation 'com.chartboost:chartboost-mediation-adapter-digital-turbine-exchange:+'
-    implementation 'com.chartboost:chartboost-mediation-adapter-google-bidding:+'
-    implementation 'com.chartboost:chartboost-mediation-adapter-inmobi:+'
-    implementation 'com.chartboost:chartboost-mediation-adapter-ironsource:+'
-    implementation 'com.chartboost:chartboost-mediation-adapter-meta-audience-network:+'
-    implementation 'com.chartboost:chartboost-mediation-adapter-mintegral:+'
-    implementation 'com.chartboost:chartboost-mediation-adapter-tapjoy:+'
-    implementation 'com.chartboost:chartboost-mediation-adapter-pangle:+'
-    implementation 'com.chartboost:chartboost-mediation-adapter-reference:+'
-    implementation 'com.chartboost:chartboost-mediation-adapter-unity-ads:+'
-    implementation 'com.chartboost:chartboost-mediation-adapter-vungle:+'
-    implementation 'com.chartboost:chartboost-mediation-adapter-yahoo:+'
-
-    // Google Play Services
-    implementation 'com.google.android.gms:play-services-ads-identifier:18.0.1'
-    implementation 'com.google.android.gms:play-services-base:18.1.0'
-
-    // android app
-    implementation 'com.google.android.material:material:1.8.0'
-    implementation 'androidx.compose.material3:material3:1.1.0-alpha06'
-    implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
-    implementation 'androidx.appcompat:appcompat:1.6.1'
-
-    // Partner Adapter SDKs
-
-    //Tapjoy
-    implementation 'com.tapjoy:tapjoy-android-sdk:12.9.1'
-
-    //Facebook
-    implementation 'com.facebook.android:audience-network-sdk:6.12.0'
-
-    //AdColony
-    implementation 'com.adcolony:sdk:4.8.0'
-
-    //AdMob
-    implementation 'com.google.android.gms:play-services-ads:21.3.0'
-
-    //Vungle
-    implementation 'com.vungle:publisher-sdk-android:6.12.0'
-    implementation 'androidx.core:core-ktx:1.9.0'
-    implementation "androidx.localbroadcastmanager:localbroadcastmanager:1.1.0"
-
-    //AppLovin
-    implementation 'com.applovin:applovin-sdk:11.5.5'
-
-    //Unity Ads
-    implementation 'com.unity3d.ads:unity-ads:4.4.1'
-
-    //ironSource
-    implementation 'com.ironsource.sdk:mediationsdk:7.2.5'
-
-    //Fyber
-    implementation 'com.fyber:marketplace-sdk:8.2.1'
-
-    //InMobi
-    implementation 'com.inmobi.monetization:inmobi-ads:10.1.1'
-
-    //Mintegral
-    implementation 'com.mbridge.msdk.oversea:videojs:16.0.31'
-    implementation 'com.mbridge.msdk.oversea:mbjscommon:16.0.31'
-    implementation 'com.mbridge.msdk.oversea:playercommon:16.0.31'
-    implementation 'com.mbridge.msdk.oversea:reward:16.0.31'
-    implementation 'com.mbridge.msdk.oversea:videocommon:16.0.31'
-    implementation 'com.mbridge.msdk.oversea:same:16.0.31'
-    implementation 'com.mbridge.msdk.oversea:interstitialvideo:16.0.31'
-    implementation 'com.mbridge.msdk.oversea:mbnative:16.0.31'
-    implementation 'com.mbridge.msdk.oversea:nativeex:16.0.31'
-    implementation 'com.mbridge.msdk.oversea:mbbanner:16.0.31'
-    implementation 'com.mbridge.msdk.oversea:mbbid:16.0.31'
-
-    //Yahoo
-    implementation 'com.verizon.ads:android-vas-standard-edition:1.14.0'
-}
