--- HeliumExample/src/main/res/drawable/green_button.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <!-- Disable background -->
-    <item android:color="@color/slate40" android:state_enabled="false" />
-
-    <!-- Enabled background -->
-    <item android:color="@color/green80" />
-</selector>
