--- HeliumExample/src/main/res/values/colors.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <color name="navy40">#303842</color>
-    <color name="green80">#8ac73b</color>
-    <color name="slate40">#c1d1e0</color>
-</resources>
