--- HeliumExample/src/main/res/values/strings.xml
@@ -1,21 +1,0 @@
-<resources>
-    <!-- Helium App Credentials-->
-    <string name="appID">5ce82fbffde3570afb4647bc</string>
-    <string name="appSignature">fd8002155dcc6f161e41a62bce49250141b1876b</string>
-
-    <!-- Helium App Placement -->
-    <string name="heliumInterstitial">CBInterstitial</string>
-    <string name="heliumRewarded">CBRewarded</string>
-    <string name="heliumBanner">CBBanner</string>
-
-    <!-- SampleApp Values -->
-    <string name="app_name">Helium Example</string>
-    <string name="str_load">LOAD</string>
-    <string name="str_show">SHOW</string>
-    <string name="str_interstitial_ad">Interstitial Ad</string>
-    <string name="str_rewarded_ad">Rewarded Ad</string>
-    <string name="log_clean">Clean Log</string>
-    <string name="str_config">Configuration</string>
-    <string name="str_config_request">Config Request</string>
-    <string name="title_activity_splash">SplashActivity</string>
-</resources>
