--- HeliumExample/src/main/res/values/styles.xml
@@ -1,16 +1,0 @@
-<resources>
-
-    <!-- Base application theme. -->
-    <style name="AppTheme" parent="Theme.AppCompat.Light.DarkActionBar">
-        <!-- Customize your theme here. -->
-        <item name="colorPrimary">@color/slate40</item>
-        <item name="colorPrimaryDark">@color/navy40</item>
-        <item name="colorAccent">@color/green80</item>
-    </style>
-
-    <style name="GreenButton" parent="Theme.AppCompat.Light.DarkActionBar">
-        <item name="colorButtonNormal">@drawable/green_button</item>
-        <item name="android:textColor">@android:color/white</item>
-    </style>
-
-</resources>
