--- HeliumExample/src/main/res/values/theme.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-
-    <style name="MainTheme" parent="Theme.AppCompat.Light.NoActionBar">
-        <item name="android:typeface">normal</item>
-        <item name="android:textSize">15sp</item>
-        <item name="android:textColor">#1a222c</item>
-    </style>
-</resources>
