+++ build.gradle
@@ -14,7 +14,7 @@ allprojects {
         mavenCentral()
         maven {
             name "Helium's maven repo"
-            url 'https://cboost.jfrog.io/artifactory/helium'
+            url 'https://cboost.jfrog.io/artifactory/chartboost-mediation'
         }
         maven {
             name "ironSource's maven repo"
@@ -35,14 +35,6 @@ allprojects {
 
         maven {
             url "https://artifact.bytedance.com/repository/pangle"
-        }
-
-        maven {
-            url "https://cboost.jfrog.io/artifactory/private-chartboost-mediation/"
-            credentials {
-                username = System.getenv("JFROG_USER")
-                password = System.getenv("JFROG_PASS")
-            }
         }
     }
 }
