+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':HeliumExample'
+include ':ChartboostMediationExample'
